// client/src/api/model/userDto.ts
/**
 * Base User API
 * Basic User API for other APIs
 *
 * The version of the OpenAPI document: 1.0
 * Contact: nikolaus@trixner.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDto { 
    id?: string;
    username?: string;
    email?: string;
}